requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
docopt==0.6.2
num2words==0.5.10
pytz==2022.1
sqlparse==0.4.2
